[Decomp][Player] Cohesion split: extract reward methods to PlayerReward.cpp
