+++ src/game/Object/Player.cpp
@@ -5904,168 +5904,10 @@ uint32 Player::GetResurrectionSpellId()
     return spell_id;
 }
 
-// Used in triggers for check "Only to targets that grant experience or honor" req
-bool Player::isHonorOrXPTarget(Unit* pVictim) const
-{
-    uint32 v_level = pVictim->getLevel();
-    uint32 k_grey  = MaNGOS::XP::GetGrayLevel(getLevel());
 
-    // Victim level less gray level
-    if (v_level <= k_grey)
-    {
-        return false;
-    }
 
-    if (pVictim->GetTypeId() == TYPEID_UNIT)
-    {
-        Creature* pVictimAsCreature = reinterpret_cast<Creature*>(pVictim);
-        if (pVictimAsCreature->IsTotem() ||
-            pVictimAsCreature->IsPet() ||
-            pVictimAsCreature->GetCreatureInfo()->ExtraFlags & CREATURE_FLAG_EXTRA_NO_XP_AT_KILL)
-        {
-            return false;
-        }
-    }
-    return true;
-}
 
-/**
- * @brief Rewards the player for killing a unit outside group reward distribution.
- *
- * @param pVictim The killed unit.
- */
-void Player::RewardSinglePlayerAtKill(Unit* pVictim)
-{
-    bool PvP = pVictim->IsCharmedOwnedByPlayerOrPlayer();
-    uint32 xp = PvP ? 0 : MaNGOS::XP::Gain(this, pVictim);
 
-    // honor can be in PvP and !PvP (racial leader) cases
-    RewardHonor(pVictim, 1);
-
-    // xp and reputation only in !PvP case
-    if (!PvP)
-    {
-        RewardReputation(pVictim, 1);
-        GiveXP(xp, pVictim);
-
-        if (Pet* pet = GetPet())
-        {
-            pet->GivePetXP(xp);
-        }
-
-        // normal creature (not pet/etc) can be only in !PvP case
-        if (pVictim->GetTypeId() == TYPEID_UNIT)
-            if (CreatureInfo const* normalInfo = ObjectMgr::GetCreatureTemplate(pVictim->GetEntry()))
-            {
-                KilledMonster(normalInfo, pVictim->GetObjectGuid());
-            }
-    }
-}
-
-/**
- * @brief Grants event kill credit to the player or nearby group members.
- *
- * @param creature_id The credited creature entry identifier.
- * @param pRewardSource The world object used for distance checks.
- */
-void Player::RewardPlayerAndGroupAtEvent(uint32 creature_id, WorldObject* pRewardSource)
-{
-    MANGOS_ASSERT((!GetGroup() || pRewardSource) && "Player::RewardPlayerAndGroupAtEvent called for Group-Case but no source for range searching provided");
-
-    ObjectGuid creature_guid = pRewardSource && pRewardSource->GetTypeId() == TYPEID_UNIT ? pRewardSource->GetObjectGuid() : ObjectGuid();
-
-    // prepare data for near group iteration
-    if (Group* pGroup = GetGroup())
-    {
-        for (GroupReference* itr = pGroup->GetFirstMember(); itr != NULL; itr = itr->next())
-        {
-            Player* pGroupGuy = itr->getSource();
-            if (!pGroupGuy)
-            {
-                continue;
-            }
-
-            if (!pGroupGuy->IsAtGroupRewardDistance(pRewardSource))
-            {
-                continue;                                    // member (alive or dead) or his corpse at req. distance
-            }
-
-            // quest objectives updated only for alive group member or dead but with not released body
-            if (pGroupGuy->IsAlive() || !pGroupGuy->HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_GHOST))
-            {
-                pGroupGuy->KilledMonsterCredit(creature_id, creature_guid);
-            }
-        }
-    }
-    else                                                    // if (!pGroup)
-    {
-        KilledMonsterCredit(creature_id, creature_guid);
-    }
-}
-
-/**
- * @brief Grants quest cast credit to the player or nearby group members.
- *
- * @param pRewardSource The credited creature or gameobject.
- * @param spellid The spell that granted the credit.
- */
-void Player::RewardPlayerAndGroupAtCast(WorldObject* pRewardSource, uint32 spellid)
-{
-    // prepare data for near group iteration
-    if (Group* pGroup = GetGroup())
-    {
-        for (GroupReference* itr = pGroup->GetFirstMember(); itr != NULL; itr = itr->next())
-        {
-            Player* pGroupGuy = itr->getSource();
-            if (!pGroupGuy)
-            {
-                continue;
-            }
-
-            if (!pGroupGuy->IsAtGroupRewardDistance(pRewardSource))
-            {
-                continue;                                // member (alive or dead) or his corpse at req. distance
-            }
-
-            // quest objectives updated only for alive group member or dead but with not released body
-            if (pGroupGuy->IsAlive() || !pGroupGuy->HasFlag(PLAYER_FLAGS, PLAYER_FLAGS_GHOST))
-            {
-                pGroupGuy->CastedCreatureOrGO(pRewardSource->GetEntry(), pRewardSource->GetObjectGuid(), spellid, pGroupGuy == this);
-            }
-        }
-    }
-    else                                                    // if (!pGroup)
-    {
-        CastedCreatureOrGO(pRewardSource->GetEntry(), pRewardSource->GetObjectGuid(), spellid);
-    }
-}
-
-/**
- * @brief Checks whether the player or corpse is close enough for shared rewards.
- *
- * @param pRewardSource The source object used for the distance check.
- * @return True if the player qualifies for group reward range; otherwise, false.
- */
-bool Player::IsAtGroupRewardDistance(WorldObject const* pRewardSource) const
-{
-    if (pRewardSource->IsWithinDistInMap(this, sWorld.getConfig(CONFIG_FLOAT_GROUP_XP_DISTANCE)))
-    {
-        return true;
-    }
-
-    if (IsAlive())
-    {
-        return false;
-    }
-
-    Corpse* corpse = GetCorpse();
-    if (!corpse)
-    {
-        return false;
-    }
-
-    return pRewardSource->IsWithinDistInMap(corpse, sWorld.getConfig(CONFIG_FLOAT_GROUP_XP_DISTANCE));
-}
 
 /**
  * @brief Gets the player's base weapon skill for an attack type.
